Moved tile, meld, etc GADT, existential and phantom to Experimental.
In internal, Tile is now a simpler version without the use of GADTs, ExistentialQuantification, phantom types, and Rank2Types.  This version is moved to Experimental.  Instead, we revert back to a simpler version and just throw error "msg" when the trying to make a non suit-tiled chow, or pung / kong / eyes with bonus tiles.  Though about using Maybe or Either for this but nah, putting that into a list would be gross, at least until I'm well versed in monad transformers.